+++ mock_phone.go
@@ -188,13 +188,7 @@ func (p *MockPhone) wireCallCallbacks(c *MockCall) {
 }
 
 func (p *MockPhone) AttendedTransfer(callA Call, callB Call) error {
-	if s := callA.State(); s != StateActive && s != StateOnHold {
-		return ErrInvalidState
-	}
-	if s := callB.State(); s != StateActive && s != StateOnHold {
-		return ErrInvalidState
-	}
-	return nil
+	return callA.AttendedTransfer(callB)
 }
 
 func (p *MockPhone) Calls() []Call {